added error handling for wonky alert definitions
Older versions of SLO/R used an array of Strings to convey the alerts we wanted to measure an SLO against. If one of these is lying in the nerdstore it will bork the new SLO calculation. This validates the inbound alerts better and returns viable NRQL so the SLO will process.

Also logs a statement with warnings to delete the SLO and recreate.