Fix for issues 78 and 64
The SLO Overview Launcher now depends on entities being set with a slor=true tag - this can be done my resaving existing SLOs. Was going to provide an automated upgrade but the tag upgrade failed inconsistently on large entity sets - it bit the bullet and require the users to resave the SLO definitions.